Added pictures of Task 4
I ordered them in the order I took them. I think the first two correspond to e-0, then the others are e-2 and e-3. But unfortunately I'm not sure abou this. Could be as well (1e-1,1e-0,1e-2). Does anyone know in which order we took the pictures? I forgot to name them accordingly :/